:recycle: refactor: DTO 디렉토리 리팩터링
