+++ src/main/java/gift/dto/betweenClient/JwtDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.betweenClient;
 
 public record JwtDTO(String token) {
