+++ src/main/java/gift/dto/betweenClient/ResponseDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.betweenClient;
 
 public record ResponseDTO(boolean isError, String message) {
