+++ src/main/java/gift/dto/betweenClient/option/OptionRequestDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.betweenClient.option;
 
 import gift.entity.Option;
 import gift.entity.Product;
