+++ src/main/java/gift/dto/betweenClient/option/OptionResponseDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.betweenClient.option;
 
 import gift.entity.Option;
