+++ src/main/java/gift/dto/betweenClient/product/ProductPostRequestDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.betweenClient.product;
 
 import gift.entity.Category;
 import gift.entity.Product;
@@ -10,7 +10,7 @@ import jakarta.validation.constraints.Pattern;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
 
-public record ProductRequestDTO(
+public record ProductPostRequestDTO(
         Long id,
 
         @NotBlank(message = "상품 이름을 입력해주세요.")
@@ -30,7 +30,17 @@ public record ProductRequestDTO(
 
         @NotBlank(message = "카테고리 이름을 입력해주세요.")
         @Length(min = 1, max = 15, message = "카테고리명 길이는 1~15자만 가능합니다.")
-        String categoryName
+        String categoryName,
+
+        @NotBlank(message = "옵션 이름을 입력해주세요.")
+        @Pattern(regexp = "[a-zA-Z0-9ㄱ-ㅎ가-힣()\\[\\]+\\-&/_ ]+", message = "( ), [ ], +, -, &, /, _을 제외한 특수문자는 입력할 수 없습니다.")
+        @Length(min = 1, max = 50, message = "옵션 이름 길이는 1~50자만 가능합니다.")
+        String optionName,
+
+        @NotNull(message = "옵션 수량을 입력해주세요.")
+        @Min(value = 1, message = "옵션 수량은 1개 이상, 1억개 미만만 가능합니다.")
+        @Max(value = 99999999, message = "옵션 수량은 1개 이상, 1억개 미만만 가능합니다.")
+        Integer optionQuantity
 ) {
     public Product convertToProduct(Category category) {
         return new Product(this.name, this.price, this.imageUrl, category);
