+++ src/main/java/gift/dto/betweenClient/wish/WishDTO.java
@@ -1,5 +1,6 @@
-package gift.dto;
+package gift.dto.betweenClient.wish;
 
+import gift.dto.betweenClient.product.ProductResponseDTO;
 import gift.entity.Wish;
 
 public record WishDTO(ProductResponseDTO productResponseDTO, Integer quantity) {
